refactor validator per route
